Basic classes structure done and Transporters partially implemented
The classes structure has been construct. The interfaces of the different modules are on Contracts. Also, there are some abstract classes to set the common functions between all the modules of the same interface.

It has been added the BaseTransporter that have common functions of Transporters.

There are two "final" implementations to get data from "uptime" and apache "acces.log". There is a fetch control, but it needs a Database (not implemented).